documents list as component
