+++ packages/frontend/src/components/Documents.tsx
@@ -1,12 +1,10 @@
 import React, { FC, useState } from "react";
 import { useQuery, useMutation, useApolloClient } from "@apollo/react-hooks";
-
 import styled from "@emotion/styled";
 import {
   Button,
   colors,
   Icon,
-  DialogModal,
   DropDown,
   Select,
   Spinner,
@@ -17,7 +15,6 @@ import { navigate } from "gatsby";
 import { Subscription } from "react-apollo";
 import { documentTypes } from "../config";
 import AppHeader from "./AppHeader";
-import DocumentCard from "./DocumentCard";
 import NewDocumentDropDown from "./NewDocumentDropDown";
 import GraphQLErrorMessage from "./GraphQLErrorMessage";
 import useUserData from "../lib/useUserData";
@@ -29,21 +26,14 @@ import {
 } from "../util";
 import {
   CREATE_DOCUMENT_MUTATION,
-  DELETE_DOCUMENT_MUTATION,
-  UPDATE_DOCUMENT_MUTATION,
   DOCUMENT_UPDATED_SUBSCRIPTION,
   EXERCISE_BY_CODE_QUERY,
-  ME_QUERY,
   CREATE_FOLDER_MUTATION,
-  FOLDER_QUERY,
-  DELETE_FOLDER_MUTATION,
-  UPDATE_FOLDER_MUTATION
+  FOLDER_QUERY
 } from "../apollo/queries";
-
 import NewExerciseButton from "./NewExerciseButton";
 import EditTitleModal from "./EditTitleModal";
-import FolderCard from "./FolderCard";
-import DocumentCardMenu from "./DocumentCardMenu";
+import DocumentListComp from "./DocumentsList";
 
 enum OrderType {
   Creation = "creation",
@@ -84,18 +74,7 @@ const Documents: FC = () => {
 
   const [order, setOrder] = useState(OrderType.Creation);
   const [searchText, setSearchText] = useState("");
-  const [deleteDocumentId, setDeleteDocumentId] = useState("");
-  const [deleteFolderId, setDeleteFolderId] = useState("");
   const [folderTitleModal, setFolderTitleModal] = useState(false);
-  const [editDocTitleModal, setEditDocTitleModal] = useState({
-    id: null,
-    title: null
-  });
-  const [editFolderNameModal, setEditFolderNameModal] = useState({
-    id: null,
-    name: null
-  });
-  const [menuOpenId, setMenuOpenId] = useState("");
   const [currentLocation, setCurrentLocation] = useState(
     userData ? userData.rootFolder : ""
   );
@@ -103,13 +82,7 @@ const Documents: FC = () => {
   let openFile = React.createRef<HTMLInputElement>();
 
   const [createDocument] = useMutation(CREATE_DOCUMENT_MUTATION);
-  const [deleteDocument] = useMutation(DELETE_DOCUMENT_MUTATION);
-  const [updateDocument] = useMutation(UPDATE_DOCUMENT_MUTATION);
-
   const [createFolder] = useMutation(CREATE_FOLDER_MUTATION);
-  const [deleteFolder] = useMutation(DELETE_FOLDER_MUTATION);
-  const [updateFolder] = useMutation(UPDATE_FOLDER_MUTATION);
-
   const {
     data: dataPage,
     loading: loadingPage,
@@ -123,9 +96,10 @@ const Documents: FC = () => {
 
   const [loadingExercise, setLoadingExercise] = useState(false);
   const [exerciseError, setExerciseError] = useState(false);
-  const { data: dataEx, loading: loadEx, error: errorEx } = useQuery(
-    EXERCISE_BY_CODE_QUERY
-  );
+
+  const onFolderClick = async (e, folder) => {
+    setCurrentLocation(folder.id);
+  };
 
   const onCreateFolder = async folderName => {
     await createFolder({
@@ -144,46 +118,6 @@ const Documents: FC = () => {
     setFolderTitleModal(false);
   };
 
-  const onUpdateDocTitle = async docTitle => {
-    await updateDocument({
-      variables: { id: editDocTitleModal.id, title: docTitle },
-      refetchQueries: [
-        {
-          query: FOLDER_QUERY,
-          variables: {
-            id: currentLocation
-          }
-        }
-      ]
-    });
-    setEditDocTitleModal({ id: null, title: null });
-    setMenuOpenId(null);
-  };
-
-  const onUpdateFolderName = async folderName => {
-    await updateFolder({
-      variables: { id: editFolderNameModal.id, input: { name: folderName } },
-      refetchQueries: [
-        {
-          query: FOLDER_QUERY,
-          variables: {
-            id: currentLocation
-          }
-        }
-      ]
-    });
-    setEditFolderNameModal({ id: null, title: null });
-    setMenuOpenId(null);
-  };
-
-  const onDocumentClick = ({ id, type }) => {
-    window.open(`/app/document/${id}`);
-  };
-
-  const onFolderClick = ({ id, type }) => {
-    setCurrentLocation(id);
-  };
-
   const onNewDocument = type => {
     window.open(`/app/document/${type}/new`);
   };
@@ -198,47 +132,6 @@ const Documents: FC = () => {
 
   const onOpenDocumentClick = () => {
     openFile.current.click();
-  };
-
-  const onDocumentMenuClick = (e, document) => {
-    e.stopPropagation();
-    if (menuOpenId === document.id) {
-      setMenuOpenId("");
-    } else {
-      setMenuOpenId(document.id);
-    }
-  };
-
-  const onDocumentRenameClick = (e, document) => {
-    e.stopPropagation();
-    setEditDocTitleModal({ id: document.id, title: document.title });
-  };
-
-  const onDocumentDeleteClick = (e, document) => {
-    e.stopPropagation();
-    setDeleteDocumentId(document.id);
-  };
-
-  const onDeleteDocument = async () => {
-    await deleteDocument({ variables: { id: deleteDocumentId } });
-    setDeleteDocumentId(null);
-    refetchPage();
-  };
-
-  const onFolderRenameClick = (e, folder) => {
-    e.stopPropagation();
-    setEditFolderNameModal({ id: folder.id, name: folder.name });
-  };
-
-  const onFolderDeleteClick = (e, folder) => {
-    e.stopPropagation();
-    setDeleteFolderId(folder.id);
-  };
-
-  const onDeleteFolder = async () => {
-    await deleteFolder({ variables: { id: deleteFolderId } });
-    setDeleteFolderId(null);
-    refetchPage();
   };
 
   const onOpenExercise = async exerciseCode => {
@@ -327,7 +220,13 @@ const Documents: FC = () => {
           {currentLocation === userData.rootFolder ? (
             <h1>Mis documentos</h1>
           ) : (
-            <h1>Mis documentos > {dataPage.folder.name}</h1>
+            <h1>
+              <a onClick={() => setCurrentLocation(userData.rootFolder)}>
+                Mis documentos
+              </a>{" "}
+              > <Icon name="folder-icon" />
+              {dataPage.folder.name}
+            </h1>
           )}
         </Header>
         <Rule />
@@ -356,7 +255,8 @@ const Documents: FC = () => {
                   setFolderTitleModal(true);
                 }}
               >
-                <Icon name="new-folder" /> Nueva carpeta
+                <Icon name="new-folder" />
+                Nueva carpeta
               </NewFolderButton>
               <NewExerciseButton
                 onOpenExercise={onOpenExercise}
@@ -382,116 +282,29 @@ const Documents: FC = () => {
             </HeaderButtons>
           </DocumentListHeader>
         )}
-
         {(documents || folders) &&
         (documents.length > 0 || folders.length > 0) ? (
           searchText ? (
             filterDocuments(documents).length > 0 ||
             filterFolders(folders).length > 0 ? (
-              <DocumentList>
-                {filterDocuments(documents).map((document: any) => (
-                  <StyledDocumentCard
-                    key={document.id}
-                    document={document}
-                    onClick={() => onDocumentClick(document)}
-                  >
-                    <DocumentMenuButton
-                      onClick={e => {
-                        onDocumentMenuClick(e, document);
-                      }}
-                    >
-                      <Icon name="ellipsis" />
-                    </DocumentMenuButton>
-                    {menuOpenId === document.id && (
-                      <DocumentCardMenu
-                        document
-                        onDelete={e => onDocumentDeleteClick(e, document)}
-                        onRename={e => onDocumentRenameClick(e, document)}
-                      />
-                    )}
-                  </StyledDocumentCard>
-                ))}
-                {filterFolders(folders).map((folder: any) => (
-                  <StyledFolderCard
-                    key={folder.id}
-                    folder={folder}
-                    onClick={() => onFolderClick(folder)}
-                  >
-                    <DocumentMenuButton
-                      onClick={e => {
-                        onDocumentMenuClick(e, folder);
-                      }}
-                    >
-                      <Icon name="ellipsis" />
-                    </DocumentMenuButton>
-                    {menuOpenId === folder.id && (
-                      <DocumentCardMenu
-                        folder
-                        onDelete={e => onFolderDeleteClick(e, folder)}
-                        onRename={e => {
-                          onFolderRenameClick(e, folder);
-                        }}
-                      />
-                    )}
-                  </StyledFolderCard>
-                ))}
-              </DocumentList>
+              <DocumentListComp
+                documents={filterDocuments(documents)}
+                folders={filterFolders(folders)}
+                currentLocation={currentLocation}
+                onFolderClick={onFolderClick}
+              />
             ) : (
               <NoDocuments>
                 <h1>No hay resultados para tu búsqueda</h1>
               </NoDocuments>
             )
           ) : (
-            <DocumentList>
-              {filterDocuments(documents).map((document: any) => (
-                <StyledDocumentCard
-                  key={document.id}
-                  document={document}
-                  onClick={() => onDocumentClick(document)}
-                >
-                  <DocumentMenuButton
-                    onClick={e => {
-                      onDocumentMenuClick(e, document);
-                      //onDocumentDeleteClick(e, document)
-                    }}
-                  >
-                    <Icon name="ellipsis" />
-                  </DocumentMenuButton>
-                  {menuOpenId === document.id && (
-                    <DocumentCardMenu
-                      document
-                      onDelete={e => onDocumentDeleteClick(e, document)}
-                      onRename={e => onDocumentRenameClick(e, document)}
-                    />
-                  )}
-                </StyledDocumentCard>
-              ))}
-              {filterFolders(folders).map((folder: any) => (
-                <StyledFolderCard
-                  key={folder.id}
-                  folder={folder}
-                  onClick={() => onFolderClick(folder)}
-                >
-                  <DocumentMenuButton
-                    onClick={e => {
-                      onDocumentMenuClick(e, folder);
-                      //onFolderDeleteClick(e, folder)
-                    }}
-                  >
-                    <Icon name="ellipsis" />
-                  </DocumentMenuButton>
-                  {menuOpenId === folder.id && (
-                    <DocumentCardMenu
-                      folder
-                      onDelete={e => onFolderDeleteClick(e, folder)}
-                      onRename={e => {
-                        onFolderRenameClick(e, folder);
-                      }}
-                    />
-                  )}
-                </StyledFolderCard>
-              ))}
-            </DocumentList>
+            <DocumentListComp
+              documents={filterDocuments(documents)}
+              folders={filterFolders(folders)}
+              currentLocation={currentLocation}
+              onFolderClick={onFolderClick}
+            />
           )
         ) : (
           <NoDocuments>
@@ -509,24 +322,6 @@ const Documents: FC = () => {
           }}
         />
       </Content>
-      <DialogModal
-        isOpen={!!deleteDocumentId}
-        title="Eliminar"
-        text="¿Seguro que quieres eliminar este documento?"
-        okText="Aceptar"
-        cancelText="Cancelar"
-        onOk={onDeleteDocument}
-        onCancel={() => setDeleteDocumentId(null)}
-      />
-      <DialogModal
-        isOpen={!!deleteFolderId}
-        title="Eliminar"
-        text="¿Seguro que quieres eliminar esta carpeta?"
-        okText="Aceptar"
-        cancelText="Cancelar"
-        onOk={onDeleteFolder}
-        onCancel={() => setDeleteFolderId(null)}
-      />
       <input
         ref={openFile}
         type="file"
@@ -542,28 +337,6 @@ const Documents: FC = () => {
           modalText="Nombre de la carpeta"
           placeholder="Carpeta sin título"
           saveButton="Crear"
-        />
-      )}
-      {editDocTitleModal.id && (
-        <EditTitleModal
-          title={editDocTitleModal.title}
-          onCancel={() => setEditDocTitleModal({ id: null, title: null })}
-          onSave={onUpdateDocTitle}
-          modalTitle="Cambiar nombre del documento"
-          modalText="Nombre del documento"
-          placeholder={editDocTitleModal.title}
-          saveButton="Cambiar"
-        />
-      )}
-      {editFolderNameModal.id && (
-        <EditTitleModal
-          title={editFolderNameModal.name}
-          onCancel={() => setEditFolderNameModal({ id: null, name: null })}
-          onSave={onUpdateFolderName}
-          modalTitle="Cambiar nombre de la carpeta"
-          modalText="Nombre de la carpeta"
-          placeholder={editFolderNameModal.name}
-          saveButton="Cambiar"
         />
       )}
     </Container>
@@ -599,6 +372,9 @@ const Header = styled.div`
     flex: 1;
     font-weight: bold;
     font-size: 24px;
+    &:hover {
+      cursor: pointer;
+    }
   }
 `;
 
@@ -622,15 +398,6 @@ const HeaderButtons = styled.div`
   justify-content: flex-end;
 `;
 
-const HeaderButton = styled(Button)`
-  padding: 0px 20px;
-  svg {
-    width: 20px;
-    height: 20px;
-    margin-right: 6px;
-  }
-`;
-
 const ViewOptions = styled.div`
   margin-right: 10px;
 `;
@@ -642,72 +409,6 @@ const OrderSelect: Select = styled(Select)`
 const SearchInput: Input = styled(Input)`
   width: 210px;
   flex: inherit;
-`;
-
-const DocumentList = styled.div`
-  display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(240px, 1fr));
-  grid-auto-rows: 1fr;
-  grid-column-gap: 40px;
-  grid-row-gap: 40px;
-  margin-bottom: 60px;
-
-  &::before {
-    content: "";
-    width: 0px;
-    padding-bottom: 85.7%;
-    grid-row: 1 / 1;
-    grid-column: 1 / 1;
-  }
-
-  & > div:first-of-type {
-    grid-row: 1 / 1;
-    grid-column: 1 / 1;
-  }
-`;
-
-const DocumentMenuButton = styled.div<{ isOpen: boolean }>`
-  position: absolute;
-  right: 14px;
-  top: 14px;
-  width: 34px;
-  height: 34px;
-  align-items: center;
-  justify-content: center;
-  border-radius: 4px;
-  border: 1px solid ${colors.gray3};
-  background-color: white;
-  display: none;
-
-  &:hover {
-    background-color: ${colors.gray1};
-    border-color: ${colors.gray4};
-  }
-
-  ${props =>
-    props.isOpen &&
-    css`
-      border: solid 1px #dddddd;
-      background-color: #e8e8e8;
-    `} svg {
-    transform: rotate(90deg);
-  }
-`;
-
-const StyledDocumentCard = styled(DocumentCard)`
-  &:hover {
-    ${DocumentMenuButton} {
-      display: flex;
-    }
-  }
-`;
-
-const StyledFolderCard = styled(FolderCard)`
-  &:hover {
-    ${DocumentMenuButton} {
-      display: flex;
-    }
-  }
 `;
 
 interface NewDocumentButtonProps {
