+++ packages/ui/src/components/icons/Folder.js
@@ -1,0 +1,15 @@
+import React from "react";
+
+const SvgFolder = props => (
+
+<svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" viewBox="0 0 20 20"{...props}>
+    <g       fill="currentColor"
+      fillRule="evenodd">
+        <path fill="#FBFBFB" d="M-188-94h1280v800H-188z"/>
+        <path fill="#373B44" fillRule="nonzero" d="M1.277 17.053h17.446V5.702h-8.145a.5.5 0 0 1-.383-.18l-2.01-2.4a.5.5 0 0 0-.382-.179H1.277v14.11zM20 4.925V17.83a.5.5 0 0 1-.5.5H.5a.5.5 0 0 1-.5-.5V2.167a.5.5 0 0 1 .5-.5h7.899a.5.5 0 0 1 .383.179l2.009 2.4a.5.5 0 0 0 .383.179H19.5a.5.5 0 0 1 .5.5z"/>
+    </g>
+</svg>
+);
+
+export default SvgFolder;
+
